+++ src/libaktualizr/storage/sqlstorage.h
@@ -97,7 +97,7 @@ class SQLStorage : public SQLStorageBase, public INvStorage {
 
   std::unique_ptr<StorageTargetWHandle> allocateTargetFile(const Uptane::Target& target) override;
   std::unique_ptr<StorageTargetRHandle> openTargetFile(const Uptane::Target& target) override;
-  boost::optional<std::pair<size_t, std::string>> checkTargetFile(const Uptane::Target& target) const override;
+  boost::optional<std::pair<uintmax_t, std::string>> checkTargetFile(const Uptane::Target& target) const override;
   std::vector<Uptane::Target> getTargetFiles() override;
   void removeTargetFile(const std::string& target_name) override;
   void cleanUp() override;
